Initial commit for generation of sources
This bash script for now relies on the fact that the current folder (pwd command in linux) is project base. It can be seen here for example:
./src/main/antlr4/dev/conductor/cql/cql.g4
- for path to grammar

Do we need to run something like
cd <poject_base>
to be sure?

Not sure if I need to specify JAVA_HOME. If so, don't know what to place.

This bash script file needs exec rights to be executed.
If git doesn't store file permissions, we need to run chmod.